Refactoring names and check style adjustments
