+++ MapboxAndroidDemo/src/gpservices/AndroidManifest.xml
@@ -157,6 +157,7 @@
         </activity>
 
         <service android:name="com.mapbox.services.android.telemetry.service.TelemetryService"/>
+        <service android:name=".account.AccountRetrievalService" />
 
         <activity
             android:name=".examples.basics.MapboxMapOptionActivity"
@@ -422,5 +423,23 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity"/>
         </activity>
+
+        <activity
+            android:name=".account.LandingActivity"
+            android:screenOrientation="portrait">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+                <action android:name="android.intent.action.VIEW" />
+
+                <category android:name="android.intent.category.DEFAULT" />
+                <category android:name="android.intent.category.BROWSABLE" />
+
+                <data
+                    android:host="authorize"
+                    android:scheme="mapbox-android-dev-preview" />
+            </intent-filter>
+        </activity>
+
     </application>
 </manifest>
