+++ MapboxAndroidDemo/src/main/AndroidManifest.xml
@@ -134,7 +134,7 @@
         </activity>
 
         <service android:name="com.mapbox.services.android.telemetry.service.TelemetryService" />
-        <service android:name=".AccountRetrievalService" />
+        <service android:name=".account.AccountRetrievalService" />
 
         <activity
             android:name=".examples.basics.MapboxMapOptionActivity"
@@ -396,7 +396,7 @@
                 android:value="com.mapbox.mapboxandroiddemo.MainActivity" />
         </activity>
         <activity
-            android:name=".LandingActivity"
+            android:name=".account.LandingActivity"
             android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
