+++ MapboxAndroidDemo/src/main/java/com/mapbox/mapboxandroiddemo/account/MapboxAccountRetrofitService.java
@@ -1,4 +1,4 @@
-package com.mapbox.mapboxandroiddemo;
+package com.mapbox.mapboxandroiddemo.account;
 
 import com.mapbox.mapboxandroiddemo.model.usermodel.UserResponse;
 
@@ -12,6 +12,8 @@ import retrofit2.http.Query;
  */
 
 public interface MapboxAccountRetrofitService {
+
   @GET("User/{username}")
   Call<UserResponse> getUserAccount(@Path("username") String userName, @Query("access_token") String accessToken);
+
 }
